refactor: 타입 파일 위치 조정 및 상수 private 객체로
