+++ src/components/MovieList.ts
@@ -1,4 +1,4 @@
-import { Movie } from "../main";
+import { Movie } from "../types/movie";
 
 export const IMAGE_BASE_URL = "https://image.tmdb.org/t/p/original";
