+++ src/types/movie.ts
@@ -1,0 +1,6 @@
+export interface Movie {
+  id: number;
+  poster_path: string;
+  vote_average: number;
+  title: string;
+}
