+++ tsconfig.json
@@ -22,5 +22,5 @@
     "noFallthroughCasesInSwitch": true,
     "noUncheckedSideEffectImports": true
   },
-  "include": ["src", "types", "vite-env.d.ts"]
+  "include": ["src", "src/types", "vite-env.d.ts"]
 }
